requirements.txt
transformers #huggingface transformers library
transformers[sentencepiece]
datasets
sacrebleu #for evaluation metrics of text summarization
rouge_score 
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
torch
notebook
boto3
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
-e . #this will tell pip to install the package in the current directory, as we used '.'